+++ components/templates/layout-main/layout-main.tsx
@@ -1,7 +1,6 @@
-import React, {useState,useCallback} from "react"
+import React from "react"
 import * as S from "./layout-main.styled"
 import {HeaderMain} from "components/organisms/headerMain"
-import {SidebarSetting} from "components/organisms/sidebarSetting";
 
 export type LayoutMainProps = {
   children: React.ReactNode,
@@ -10,24 +9,9 @@ export type LayoutMainProps = {
 export const LayoutMain = ({
   children,
 }: LayoutMainProps) => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const handleOpenMenu: React.MouseEventHandler = useCallback(() => {
-    setIsMenuOpen(true);
-  },[]);
-
-  const handleCloseMenu: React.MouseEventHandler = useCallback((event) => {
-    if (!(event.target as HTMLElement).closest(".menu")) {
-      setIsMenuOpen(false);
-    }
-  },[]);
-
   return (
     <S.LayoutMain>
-      <HeaderMain onMenuOpen={handleOpenMenu}></HeaderMain>
-      <SidebarSetting
-        isMenuOpen={isMenuOpen}
-        onCloseMenu={handleCloseMenu}
-      ></SidebarSetting>    
+      <HeaderMain />   
       <S.Contents>
         {children}
       </S.Contents>
